requirements.txt
torch==1.10.0
numpy==1.20.1
numpydoc==1.1.0
transformers==4.3.3
bert-extractive-summarizer==0.7.1
scikit-learn==0.24.1
scipy==1.6.2
newspaper3k==0.2.8
nltk==3.6.1
psycopg2==2.9.1
pandas==1.2.4
beautifulsoup4==4.9.3
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
Flask==1.1.2



